Revert "Generate ManagementContextConfiguration.imports file from annotations"
This reverts commit 6b3b0dd3a6ac4818388e97e121de69a91e21657d.
