+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.actuate.autoconfigure.web.ManagementContextConfiguration.imports
@@ -1,0 +1,10 @@
+org.springframework.boot.actuate.autoconfigure.endpoint.web.ServletEndpointManagementContextConfiguration
+org.springframework.boot.actuate.autoconfigure.endpoint.web.jersey.JerseyWebEndpointManagementContextConfiguration
+org.springframework.boot.actuate.autoconfigure.endpoint.web.reactive.WebFluxEndpointManagementContextConfiguration
+org.springframework.boot.actuate.autoconfigure.endpoint.web.servlet.WebMvcEndpointManagementContextConfiguration
+org.springframework.boot.actuate.autoconfigure.security.servlet.SecurityRequestMatchersManagementContextConfiguration
+org.springframework.boot.actuate.autoconfigure.web.jersey.JerseySameManagementContextConfiguration
+org.springframework.boot.actuate.autoconfigure.web.jersey.JerseyChildManagementContextConfiguration
+org.springframework.boot.actuate.autoconfigure.web.reactive.ReactiveManagementChildContextConfiguration
+org.springframework.boot.actuate.autoconfigure.web.servlet.ServletManagementChildContextConfiguration
+org.springframework.boot.actuate.autoconfigure.web.servlet.WebMvcEndpointChildContextConfiguration
