+++ spring-boot-project/spring-boot-tools/spring-boot-autoconfigure-processor/src/main/resources/META-INF/gradle/incremental.annotation.processors
@@ -1,3 +1,2 @@
 org.springframework.boot.autoconfigureprocessor.AutoConfigureAnnotationProcessor,aggregating
 org.springframework.boot.autoconfigureprocessor.AutoConfigurationImportsAnnotationProcessor,aggregating
-org.springframework.boot.autoconfigureprocessor.ManagementContextConfigurationImportsAnnotationProcessor,aggregating
