+++ spring-boot-project/spring-boot-tools/spring-boot-autoconfigure-processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,3 +1,2 @@
 org.springframework.boot.autoconfigureprocessor.AutoConfigureAnnotationProcessor
 org.springframework.boot.autoconfigureprocessor.AutoConfigurationImportsAnnotationProcessor
-org.springframework.boot.autoconfigureprocessor.ManagementContextConfigurationImportsAnnotationProcessor
